Extra Player Instantiation Bug Fixed
I went around the bug that caused an extra player to be instantiated on a new scene load. Instead, I disabled the lobby as the room scene, and did not destroy the object on load. In the game scene, I reset all the camera and positions of the character prefabs